Extract PostStream state
