+++ js/src/forum/components/DiscussionPage.js
@@ -8,6 +8,7 @@ import SplitDropdown from '../../common/components/SplitDropdown';
 import listItems from '../../common/helpers/listItems';
 import DiscussionControls from '../utils/DiscussionControls';
 import DiscussionList from './DiscussionList';
+import PostStreamState from '../state/PostStreamState';
 
 /**
  * The `DiscussionPage` component displays a whole discussion page, including
@@ -89,6 +90,24 @@ export default class DiscussionPage extends Page {
   view() {
     const discussion = this.discussion;
 
+    let content;
+
+    if (discussion) {
+      const stream = new PostStream({ state: this.stream });
+      stream.on('positionChanged', this.positionChanged.bind(this));
+      content = [
+        DiscussionHero.component({ discussion }),
+        <div className="container">
+          <nav className="DiscussionPage-nav">
+            <ul>{listItems(this.sidebarItems().toArray())}</ul>
+          </nav>
+          <div className="DiscussionPage-stream">{stream.render()}</div>
+        </div>,
+      ];
+    } else {
+      content = LoadingIndicator.component({ className: 'LoadingIndicator--block' });
+    }
+
     return (
       <div className="DiscussionPage">
         {app.discussions.hasDiscussions() ? (
@@ -99,19 +118,7 @@ export default class DiscussionPage extends Page {
           ''
         )}
 
-        <div className="DiscussionPage-discussion">
-          {discussion
-            ? [
-                DiscussionHero.component({ discussion }),
-                <div className="container">
-                  <nav className="DiscussionPage-nav">
-                    <ul>{listItems(this.sidebarItems().toArray())}</ul>
-                  </nav>
-                  <div className="DiscussionPage-stream">{this.stream.render()}</div>
-                </div>,
-              ]
-            : LoadingIndicator.component({ className: 'LoadingIndicator--block' })}
-        </div>
+        <div className="DiscussionPage-discussion">{content}</div>
       </div>
     );
   }
@@ -197,7 +204,7 @@ export default class DiscussionPage extends Page {
     // Set up the post stream for this discussion, along with the first page of
     // posts we want to display. Tell the stream to scroll down and highlight
     // the specific post that was routed to.
-    this.stream = new PostStream({ discussion, includedPosts });
+    this.stream = new PostStreamState(discussion, includedPosts);
     this.stream.on('positionChanged', this.positionChanged.bind(this));
     this.stream.goToNumber(m.route.param('near') || (includedPosts[0] && includedPosts[0].number()), true);
 
@@ -268,7 +275,7 @@ export default class DiscussionPage extends Page {
     items.add(
       'scrubber',
       PostStreamScrubber.component({
-        stream: this.stream,
+        state: this.stream,
         className: 'App-titleControl',
       }),
       -100
